Unify optimized implementations
