+++ setup.cfg
@@ -27,91 +27,76 @@ packages =
     ethereum/utils
     ethereum/crypto
     ethereum_optimized/
-    ethereum_optimized/frontier
     ethereum/homestead
     ethereum/homestead/utils
     ethereum/homestead/vm
     ethereum/homestead/vm/instructions
     ethereum/homestead/vm/precompiled_contracts
-    ethereum_optimized/homestead
     ethereum/dao_fork
     ethereum/dao_fork/utils
     ethereum/dao_fork/vm
     ethereum/dao_fork/vm/instructions
     ethereum/dao_fork/vm/precompiled_contracts
-    ethereum_optimized/dao_fork
     ethereum/tangerine_whistle
     ethereum/tangerine_whistle/utils
     ethereum/tangerine_whistle/vm
     ethereum/tangerine_whistle/vm/instructions
     ethereum/tangerine_whistle/vm/precompiled_contracts
-    ethereum_optimized/tangerine_whistle
     ethereum/spurious_dragon
     ethereum/spurious_dragon/utils
     ethereum/spurious_dragon/vm
     ethereum/spurious_dragon/vm/instructions
     ethereum/spurious_dragon/vm/precompiled_contracts
-    ethereum_optimized/spurious_dragon
     ethereum/byzantium
     ethereum/byzantium/utils
     ethereum/byzantium/vm
     ethereum/byzantium/vm/instructions
     ethereum/byzantium/vm/precompiled_contracts
-    ethereum_optimized/byzantium
     ethereum/constantinople
     ethereum/constantinople/utils
     ethereum/constantinople/vm
     ethereum/constantinople/vm/instructions
     ethereum/constantinople/vm/precompiled_contracts
-    ethereum_optimized/constantinople
     ethereum/istanbul
     ethereum/istanbul/utils
     ethereum/istanbul/vm
     ethereum/istanbul/vm/instructions
     ethereum/istanbul/vm/precompiled_contracts
-    ethereum_optimized/istanbul
     ethereum/muir_glacier
     ethereum/muir_glacier/utils
     ethereum/muir_glacier/vm
     ethereum/muir_glacier/vm/instructions
     ethereum/muir_glacier/vm/precompiled_contracts
-    ethereum_optimized/muir_glacier
     ethereum/berlin
     ethereum/berlin/utils
     ethereum/berlin/vm
     ethereum/berlin/vm/instructions
     ethereum/berlin/vm/precompiled_contracts
-    ethereum_optimized/berlin
     ethereum/london
     ethereum/london/utils
     ethereum/london/vm
     ethereum/london/vm/instructions
     ethereum/london/vm/precompiled_contracts
-    ethereum_optimized/london
     ethereum/arrow_glacier
     ethereum/arrow_glacier/utils
     ethereum/arrow_glacier/vm
     ethereum/arrow_glacier/vm/instructions
     ethereum/arrow_glacier/vm/precompiled_contracts
-    ethereum_optimized/arrow_glacier
     ethereum/gray_glacier
     ethereum/gray_glacier/utils
     ethereum/gray_glacier/vm
     ethereum/gray_glacier/vm/instructions
     ethereum/gray_glacier/vm/precompiled_contracts
-    ethereum_optimized/gray_glacier
     ethereum/paris
     ethereum/paris/utils
     ethereum/paris/vm
     ethereum/paris/vm/instructions
     ethereum/paris/vm/precompiled_contracts
-    ethereum_optimized/paris
     ethereum/shanghai
     ethereum/shanghai/utils
     ethereum/shanghai/vm
     ethereum/shanghai/vm/instructions
     ethereum/shanghai/vm/precompiled_contracts
-    ethereum_optimized/shanghai
 
 
 package_dir =
